# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.1",
        "description": "This is my version of API made up in #BackEndChallengeAlura.",
        "title":"shareflix-api-alurachallenge",
        "contact": {
            "name": "Adriano Barbosa",
            "email": "b.adrianobarbosa@gmail.com"            
        },
        "license": "ISC"
    },
    "servers": [
        {
            "url": "http://localhost:3000/",
            "description": "local server"
        }
    ],
    "tags": [
        {
            "name": "CRUD operations"
        },
        {
            "name": "Movies"
        },
        {
            "name": "Categories"
        }
    ],
    "paths": {
        "/videos": {
            "get": {
                "tags": ["Movies"],
                "description": "Get all movies",
                "operationId": "getAllMovies",
                "parameters": [
                    {
                        "in":"query",
                        "name": "search",                    
                        "required": false,
                        "description": "Query string to find a movie by title or description"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movies were obtained",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movies"
                                }
                            }
                        }
                    }                   
                }
            },
            "post": {
                "tags": ["Movies"],
                "description": "Create movies",
                "operationId": "madeUpMovie",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Movie"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "New movie was created"
                    }
                }
            }
        },
        "/videos/:id": {
            "get":{
                "tags": ["Movies"],
                "description": "Retrieve one Movie",
                "operationId": "getOneMovie",
                "parameters": [
                    {
                        "name": "id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        },
                        "required": true,
                        "description": "Movie ID"
                    }
                ],                
                "responses": {
                    "200": {
                        "description": "Retrieve one movie",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["Movies"],
                "description": "Update movie",
                "operationId": "updateMovie",
                "parameters": [
                    {
                        "name": "id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        },
                        "required": true,
                        "description": "Movie ID"
                    }  
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Movie"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Movie updated",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Movies"],
                "description": "delete one movie",
                "operationId": "deleteMovie",
                "parameters": [
                    {
                        "name": "id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        },
                        "required": true,
                        "description": "Movie ID"
                    }  
                ],                
                "responses": {
                    "200": {                        
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categorias":{
            "get":{
                "tags": ["Categories"],
                "description": "Get all categories",
                "operationId": "getAllCategories",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Categories are obtained",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categories"
                                }
                            }
                        }
                    }                   
                }
            },
            "post":{
                "tags": ["Categories"],
                "description": "Create Category",
                "operationId": "madeUpCategory",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "New category was created"
                    }
                }
            }
        },
        "/categorias/:id":{
            "get":{
                "tags": ["Categories"],
                "description": "Retrieve one Category",
                "operationId": "getOneCategory",
                "parameters": [
                    {
                        "name": "id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        },
                        "required": true,
                        "description": "Category ID"
                    }
                ],                
                "responses": {
                    "200": {
                        "description": "Retrieve one category",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["Categories"],
                "description": "Update category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        },
                        "required": true,
                        "description": "Category ID"
                    }  
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Categories"],
                "description": "delete one category",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        },
                        "required": true,
                        "description": "category ID"
                    }  
                ],                
                "responses": {
                    "200": {                        
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categorias/:id/videos":{
            "get":{
                "tags": ["Categories"],
                "description": "Get all movies from a category",
                "operationId": "getMoviesFromCategory",
                "parameters": [
                    {
                        "name": "id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        },
                        "required": true,
                        "description": "category ID"
                    }  
                ],
                "responses": {
                    "200": {
                        "description": "Movies were obtained",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movies"
                                }
                            }
                        }
                    }                   
                }
            }
        }
    },
    "components": {
        "schemas": {
            "id": {
                "type": "integer",
                "description": "Movie ID",
                "example": 1000
            },
            "titulo": {
                "type": "string",
                "example": "Some title"
            },
            "descricao": {
                "type": "string",
                "example": "here goes the explanation about the thing"
            },
            "cor": {
                "type":"string",
                "example": "an hexadecimal like #000000"
            },
            "url": {
                "type": "string",
                "example": "https://myawesome.movie"
            },
            "Movie": {
                "type": "object",
                "properties": {
                   "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "titulo": {
                        "$ref": "#/components/schemas/titulo"
                    },
                    "descricao": {
                        "$ref": "#/components/schemas/descricao"
                    },
                    "url": {
                        "$ref": "#/components/schemas/url"
                    },
                    "category_id": {
                        "$ref": "#/components/schemas/id",
                        "description": "Category ID"
                    }
                }
            },
            "Category":{
                "type":"object",
                "properties":{
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "titulo": {
                        "$ref": "#/components/schemas/titulo"
                    },
                    "cor": {
                        "$ref": "#/components/schemas/cor"
                    }
                }
            },
            "Movies": {
                "type": "object",
                "properties": {
                    "movies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Movie"
                        }
                    }
                }
            },
            "Categories": {
                "type": "object",
                "properties": {
                    "movies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                   "message": {
                        "type": "string"
                    }
                }
            }
        }
    }
}